+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 *   An example for `invoice` payment method.
 *   Charge methods `getCancelledAmount` and `getTotalAmount`.
 *   Authorize method `getCancelledAmount`.
+*   Detailed `keypair` fetch.
 
 ### Fixed
 *   Problem with HeidelpayApiException.